fix font sizes as functions for curved labels
Each glyph has a `a_minzoom` and `a_maxzoom` which set the range of zoom
levels it is safe to show the glyph if the glyph is rendered at the
fontSize which was used for placement (the one set in the bucket).
`u_zoom` is compared against the min and max values to see if the glyph
can be shown.

When the font size in the style is different than that at which the
glyphs were placed, then the min and max zooms the glyph can be shown at
change. Instead of changing the min and max zooms, this adjusts `u_zoom`
so that less work needs to be done in the shader.