fix compositing
fixes #338

Previously, composited layers were added on to neighbouring groups
with hopes of reducing the number of clipping mask redraws.

This failed if the source of the group it is added on to didn't have
a tile for that area, because then the composited layer would not
get composited onto the main canvas.

Composited layers now form their own layerGroup and are composited in
one call, not tile by tile.