Tolerate extra keys.
This is a failing test that demonstrates the issue: it would need
a fix - specifically a removal of the unknown key behavior at the root
object - to pass.

I think we should tolerate extra keys in style objects because they're
very useful for the editor use case, like if you want to store extra
data for color transforms etc., that is not handled by the gl renderer
directly (see https://github.com/mapbox/mapbox-gl-js/pull/479 )

The only potential drawback I see here is if gl-native's parsing
behavior would have problems with additional data, but it seems
very unlikely.