Reload layer source after adding
When calling `map.addLayer()`, if geometry is being styled for the first
time, the new geometry does not render. However, as you zoom or pan the
map to a new tile the correct geometry renders.

New layers with sources need their sources to be refreshed in order to
pickup previously unused geometry.