fix blank areas in raster layers
fix #1876

This adapts raster fading for the changes in #1603. The change in the
return value meant that some tiles were marked as covered even when they
weren't.

Also, `retain` was being modified while being iterated over which could
have lead to retaining parents of parents of parents... or children of
children of children. It now iterates over only those id's present in
`retain` before the iteration begins.