use function instead of string of code for getValue
A function of the layer, globalProperties, featureProperties is simpler
than a string of code that gets joined into a function.

This separates the vertex push method for core attributes from the push
methods for data-driven attributes. This will let us:
- use a regular function for adding core attributes to buffers
- and either:
    - generate a separate method for data driven attributes
    - or optimize adding attributes without code generation

Data driven properties need to be evaluated once per feature, not once
per vertex. This means we can evaluate them outside the main vertex loop
and then add the same value for every vertex.

The hot loop for lines, fills and symbols will be:
for (var i = startIndex; i < endIndex; i++) {
    ...
}

If it's too slow in the current implementation we could add new
StructArray methods for setting a single value for a range of elements.
If we need to generate code it would be nice to contain that to
StructArray.