Use sinon's sandbox API
Create a sandbox in a global `beforeEach` callback, and restore it in a global `afterEach` callback. Inject the methods into the tap `Test` object so that they are accessible via `t.spy`, `t.stub`, etc., and use those exclusively.

This ensures that a stubbed or spied global method (e.g. `util.warnOnce`) does not linger beyond a single test case.