fix tinting rasters with alpha, fix #3723
Rasters need to be non-premultiplied when blended. This switches the
textures to non-premultiplied and the premultiplies them at the end.

I also changed how fading, mixing and opacity uniforms are calculated
and passed to the shaders. Have the `mix` value separate from the
`opacity` is necessary for us to be able to blend in unpremultiplied
space.