Indicate macOS SDK support (#641)
* Indicate macOS SDK support

Added macOS SDK support information alongside the existing iOS SDK support information. Clarified that, with respect to the iOS and macOS SDKs, this document is intended as a file format reference for style authors rather than a runtime styling reference.

* macOS SDK v0.2.1

* Removed link to iOS SDK API reference

We’ll add the link back in as soon as iOS SDK v3.4.0 and macOS SDK v0.3.0 are officially released.