add stopgap to for querying updated symbol buckets
The index relied on a source layer and index to identify a feature. If
a data layer gets updated a new bucket is formed where these identifiers
probably don't match the previous bucket. Since there is a gap between a
bucket being loaded and the index being updated (during the next
placement) it was possible for a query to return wrong results.

This prevents it from returning incorrect features but it does not
return the correct features. This might cause it to return empty
results.