Fix hash bug (#6490)
* Failing test for map#remove + hash option

This test demonstrates that a type error thrown (“Cannot read property 'getCenter' of undefined”) when initializing a map with hash option set to true, then calling map#remove. The createMap() helper is avoided because this issue appears to be specific to when initializing the map with the options object.

* Add clearTimeout to remove past timeouts on remove

This fix adds clearTimeout in the remove method to clear out any currently-running timeouts from the throttling behavior. This fixes an issue in which a type error was thrown if the map was removed while a timeout still existed.