Ignore unfamiliar render tests operations
Before this change, the `suite_implementation` treated any unfamiliar
render tests operations as `Map` interface methods and tried to invoke
them.

This PR checks if the operations can be invoked on the map object and if
not, these operations are ignored.

This change is required because gl-native is extending render-tests operations
with the API for collecting various metrics.