add gpu timing map events (#8829)
* Introduce 'gpuTiming' map options.

Originally implemented by Chris. I've rebased it and exposed it
with just the event listeners instead of a map option.

Listen to `gpu-timing-frame` to get the gpu time for the frame and
listen to `gpu-timing-layer` to get the gpu time for all individual
layers. It is not recommended to listen to both.

* fixup