fix pinching while touching markers (#9683)
* fix multi-finger gestures touching markers

Previously we were using e.targetTouches as the list of touches relevant
to the map. It contains only the touches that first touched the same
target. Touches that hit a marker have a different target.

This fixes uses e.touches and filters out touches who's target is not a
child of the map.

fix #9675

* update tests and add regression test