project/unproject uses terrain (internal-259)
Project and unproject API take into account terrain elevation.

Following API usage and questions on external forums and committed issues, didn't identify a single use case that would require project unproject in 3D mode with terrain to behave as if terrain is not active. In most of the cases, it is about getting LngLat for pointer position or for corners of screen.

Good examples for API are marker implementation and query_features.html, with terrain enabled in this patch.

Revert intermediate internal transform pointLocation3D and locationPoint3D used in unit tests and for marker development by project/unproject (as it was before internal-183). Proof that existing client code is just expected to continue working with this change.

If there is additional need for API change, we will identify from pre-release customers and add later.

Fixes: internal-196