Adjust clamping of perspective ratio for symbols and text (internal-271)
The perspective ratio value for text and symbols is currently bound between [0, 4]. This means we may allow symbols to grow to up to 4 times their original size introducing blurriness and distortion. mapbox/mapbox-gl-js#6060 was not taking into account tilt to go over 60 so we need to revisit that.

To have perfect matching between font size and symbol sizes, we would clamp that value in the [0, 1] range. Though using this range lead to labels feeling smaller than they actually are. This is due to the fact that their surroundings, which is very big as it is close to the camera, make them appear smaller due to a perspective optical illusion. The bound is now clamped to a max 1.5x to not allow any extreme bluriness and oversized iconography and text. We will revisit the perspective ratio formula for a linear interpolation from front to back in [0, 1].